# Conflicts:
#	plugins/generator-1.21.8/neoforge-1.21.8/templates/gui/gui_msg_slider.java.ftl

+++ src/main/java/net/mcreator/element/types/Dimension.java
@@ -58,7 +58,7 @@ import java.util.List;
 	public String defaultEffects;
 	public boolean useCustomEffects;
 	public boolean hasClouds;
-	public double cloudHeight;
+	public int cloudHeight;
 	public String skyType;
 	public Color airColor;
 	public boolean sunHeightAffectsFog;
